Rename files to acceptable python names
